requirements.txt
asgiref==3.5.2
attr==0.3.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
idna==3.4
Naked==0.1.32
Pillow==9.3.0
pycparser==2.21
pycryptodome==3.16.0
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
wget==3.2
